new_tax/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
websockets==14.1
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
PyYAML==6.0.2

faster-whisper==1.1.1
gTTS==2.5.4
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
langchain-huggingface==0.1.2
langchain-chroma==0.1.4
chromadb==0.5.23
sentence-transformers==3.3.1
tiktoken==0.8.0

httpx==0.28.1
